GPII-1176:  Renamed overly general `handler` and `middleware` files per pull request feedback.
