+++ index.js
@@ -1,4 +1,4 @@
 // Require all of the server-side components at once.
 require("./validator");
-require("./src/js/server/middleware");
-require("./src/js/server/handler");
+require("./src/js/server/schemaMiddleware");
+require("./src/js/server/schemaHandler");
